Add frequency calculation and minor refactoring

+++ RobotArmController/MotorControl/inc/MotorControl/speed_control.h
@@ -6,17 +6,17 @@
 
 #include <stm32f4xx_hal.h>
 
-#define TIMER_PRESCALER_MIN ((double)(1))
-#define TIMER_PRESCALER_MAX ((double)(2e16))
-#define TIMER_PERIOD_MIN ((double)(1))
-#define TIMER_PERIOD_MAX ((double)(2e16))
+#define TIMER_PRESCALER_MIN (1.0)
+#define TIMER_PRESCALER_MAX (2e16)
+#define TIMER_PERIOD_MIN (1.0)
+#define TIMER_PERIOD_MAX (2e16)
 
 // 1.8° = 0.03141593rad
-#define MOTOR_STEP_ANGLE_R ((double)0.03141593)
+#define MOTOR_STEP_ANGLE_R (0.03141593)
 // 3.5arcmin = 0.05833333° = 0.00101811rad
-#define MOTOR_STEP_ANGLE_PHI ((double)0.00101811)
+#define MOTOR_STEP_ANGLE_PHI (0.00101811)
 // 1.8° = 0.03141593rad
-#define MOTOR_STEP_ANGLE_Z ((double)0.03141593)
+#define MOTOR_STEP_ANGLE_Z (0.03141593)
 
 #define MOTOR_THEORETICAL_SPEED_MIN(_HMOTOR_) \
         ((_HMOTOR_->timer->frequency * _HMOTOR_->step_angle) \
@@ -26,11 +26,11 @@
         / (TIMER_PRESCALER_MIN * TIMER_PERIOD_MIN))
 
 // 3.95cm = 0.0395m (/rad)
-#define JOINT_TRANSFORM_FACTOR_R ((double)0.0395)
+#define JOINT_TRANSFORM_FACTOR_R (0.0395)
 // 1 (rad/rad)
-#define JOINT_TRANSFORM_FACTOR_PHI ((double)1.0)
+#define JOINT_TRANSFORM_FACTOR_PHI (1.0)
 // 2mm / (2 * pi) = 0.00031831m (/rad)
-#define JOINT_TRANSFORM_FACTOR_Z ((double)0.00031831)
+#define JOINT_TRANSFORM_FACTOR_Z (0.00031831)
 
 #define JOINT_THEORETICAL_SPEED_MIN(_HJOINT_) \
         (MOTOR_THEORETICAL_SPEED_MIN(_HJOINT_->motor) \
